added presentation 3 paging tests
