+++ test/fixtures/manifests.js
@@ -46,7 +46,8 @@ module.exports = {
     "scroll": "http://localhost:3001/scroll.json",
     "storyofwellcome": "http://localhost:3001/storyofwellcome.json",
     "tankeryshouse": "http://localhost:3001/tankeryshouse.json",
-    "theprincipleofrelativity": "http://localhost:3001/theprincipleofrelativity.json",
+    "presentation2Paging": "http://localhost:3001/presentation2-paging.json",
+    "presentation3Paging": "http://localhost:3001/presentation3-paging.json",
     "translations": "http://localhost:3001/translations.json",
     "wellcomeapocalypse": "http://localhost:3001/wellcomeapocalypse.json",
     "witnesstopeter": "http://localhost:3001/witnesstopeter.json",
